vgui: Do not use unsigned long as it is different on LP64

+++ public/vgui/IScheme.h
@@ -22,7 +22,8 @@ namespace vgui
 {
 
 typedef uintp HScheme;
-typedef unsigned long HTexture;
+// dimhotepus: unsigned long -> uint32. x86-64
+typedef uint32 HTexture;
 
 class IBorder;
 class IImage;
